# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StreamPay API",
    "version": "2.0.0",
    "description": "REST API for the StreamPay payment-streaming platform on Stellar.\n\nAll error responses use the canonical envelope:\n```json\n{ \"error\": { \"code\": \"ERROR_CODE\", \"message\": \"Human-readable detail.\", \"request_id\": \"req_...\" } }\n```\n\nv1 stream paths (`/api/streams`) are deprecated and will be removed. Use `/api/v2/streams`."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Obtain a token via POST /api/auth/wallet, then pass it as `Authorization: Bearer <token>`."
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code.",
                "example": "NOT_FOUND"
              },
              "message": {
                "type": "string",
                "description": "Human-readable description safe to display to clients.",
                "example": "The requested stream does not exist."
              },
              "request_id": {
                "type": "string",
                "description": "Opaque identifier for this request, useful for log correlation.",
                "example": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      },
      "StreamV1": {
        "type": "object",
        "description": "v1 payment stream shape. **Deprecated** \u2014 use `StreamV2` via `/api/v2/streams`.",
        "required": [
          "id",
          "recipient",
          "rate",
          "schedule",
          "status",
          "nextAction",
          "token",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique stream identifier.",
            "example": "stream-abc12345"
          },
          "recipient": {
            "type": "string",
            "description": "Stellar public key of the payment recipient.",
            "example": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG"
          },
          "rate": {
            "type": "string",
            "description": "Numeric payment amount per schedule interval.",
            "example": "120"
          },
          "schedule": {
            "type": "string",
            "enum": [
              "second",
              "minute",
              "hour",
              "day",
              "week",
              "month",
              "year"
            ],
            "description": "Payment interval.",
            "example": "month"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "active",
              "paused",
              "ended"
            ],
            "description": "Current lifecycle state of the stream."
          },
          "nextAction": {
            "type": [
              "string",
              "null"
            ],
            "description": "The next permitted action for the authenticated user.",
            "example": "start"
          },
          "token": {
            "type": "string",
            "description": "Stellar asset code.",
            "example": "XLM"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 creation timestamp.",
            "example": "2024-01-15T10:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 last-updated timestamp.",
            "example": "2024-01-15T10:00:00.000Z"
          }
        }
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad request \u2014 invalid input.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": {
                "code": "BAD_REQUEST",
                "message": "Request body must include 'recipient' and 'rate'.",
                "request_id": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": {
                "code": "INTERNAL_SERVER_ERROR",
                "message": "An unexpected error occurred.",
                "request_id": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/streams/search": {
      "get": {
        "operationId": "searchStreams",
        "summary": "Search payment streams with FTS and filters",
        "description": "Performs full-text search across stream fields (id, recipient, memo, label, email, senderAddress, partnerId, token) and filters by status, asset, sender, recipient, and creation date bounds.",
        "tags": [
          "Streams v1"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Full-text search query term.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by stream status (draft, active, paused, ended, withdrawn, cancelled).",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "active",
                "paused",
                "ended",
                "withdrawn",
                "cancelled"
              ]
            }
          },
          {
            "name": "asset",
            "in": "query",
            "description": "Filter by asset or token symbol.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sender",
            "in": "query",
            "description": "Filter by sender address.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recipient",
            "in": "query",
            "description": "Filter by recipient.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "ISO-8601 creation date lower bound.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "ISO-8601 creation date upper bound.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return (1-200, default 50).",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset (default 0).",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta",
                    "links"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/StreamV1"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "total",
                        "limit",
                        "offset",
                        "count",
                        "query"
                      ],
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "offset": {
                          "type": "integer"
                        },
                        "count": {
                          "type": "integer"
                        },
                        "query": {
                          "type": [
                            "string",
                            "null"
                          ]
                        }
                      }
                    },
                    "links": {
                      "type": "object",
                      "required": [
                        "self"
                      ],
                      "properties": {
                        "self": {
                          "type": "string"
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "results-found": {
                    "summary": "Search returned matching streams",
                    "value": {
                      "data": [
                        {
                          "id": "stream-abc12345",
                          "recipient": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG",
                          "rate": "120",
                          "schedule": "month",
                          "status": "active",
                          "nextAction": "pause",
                          "token": "XLM",
                          "createdAt": "2024-01-15T10:00:00.000Z",
                          "updatedAt": "2024-01-16T08:30:00.000Z"
                        }
                      ],
                      "meta": { "total": 1, "limit": 50, "offset": 0, "count": 1, "query": "abc12345" },
                      "links": { "self": "/api/streams/search?q=abc12345&limit=50" }
                    }
                  },
                  "no-results": {
                    "summary": "No streams matched the search query",
                    "value": {
                      "data": [],
                      "meta": { "total": 0, "limit": 50, "offset": 0, "count": 0, "query": "nonexistent" },
                      "links": { "self": "/api/streams/search?q=nonexistent&limit=50" }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
...
}